accessimed/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "accessimed"
version = "0.1.0"
description = "Open-source WCAG accessibility scanning and remediation CLI for website teams."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["accessibility", "wcag", "cli", "testing", "healthcare", "a11y"]
dependencies = [
  "aiosqlite==0.21.0",
  "anthropic==0.57.1",
  "beautifulsoup4==4.13.4",
  "fastapi==0.115.0",
  "httpx==0.28.1",
  "langgraph==0.3.14",
  "openai==2.6.1",
  "playwright==1.48.0",
  "pydantic-settings==2.10.1",
  "reportlab==4.4.0",
  "sqlalchemy==2.0.35",
  "uvicorn[standard]==0.34.0",
]

[project.urls]
Homepage = "https://github.com/ayushgawai/accessimed"
Repository = "https://github.com/ayushgawai/accessimed"
Issues = "https://github.com/ayushgawai/accessimed/issues"

[project.optional-dependencies]
dev = [
  "pytest==8.4.1",
  "pytest-asyncio==1.1.0",
  "pytest-cov==6.2.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[project.scripts]
accessimed = "app.cli:main"
